Use select to implement unblocking IO when sending broker aware requests