@yyejun 2015-10-28
overall I feel the benefit of this change is small. It barely saves the time to handle the first connection, assuming the majority time is spent in the recv call. Because eventually all requests need to be processed.. The actual receiving still in serial, after first request processed, it's very likely all other socket are all ready to be received, so the select on 2nd request, mostly like all socket are ready.