@ecanzonieri 2015-10-28
The improvements in our larger kafka cluster is notable. For watermark fetch the latency goes down from 1.70s to 0.35s, while consumer offset fetch goes down from 12.5 seconds to 1.20s.